WindowManager + addressed some comments
