+++ include/nbl/ui/IWindow.h
@@ -59,7 +59,7 @@ public:
             w->m_flags |= ECF_HIDDEN;
             onWindowHidden_impl();
         }
-        void onWindowMoved(IWindow* w, int x, int y)
+        void onWindowMoved(IWindow* w, int32_t x, int32_t y)
         {
             onWindowMoved_impl(x, y);
         }
@@ -122,7 +122,7 @@ public:
     protected:
         virtual void onWindowShown_impl() {}
         virtual void onWindowHidden_impl() {}
-        virtual void onWindowMoved_impl(int x, int y) {}
+        virtual void onWindowMoved_impl(int32_t x, int32_t y) {}
         virtual void onWindowResized_impl(uint32_t w, uint32_t h) {}
         virtual void onWindowRotated_impl() {}
         virtual void onWindowMinimized_impl() {}
