+++ include/nbl/ui/IWindowManager.h
@@ -1,0 +1,12 @@
+#ifndef I_WINDOWMANAGER
+#define I_WINDOWMANAGER
+
+namespace nbl::ui
+{
+	class IWindowManager
+	{
+	public:
+		virtual ~IWindowManager() = default;
+	};
+}
+#endif
